Small win32 compile fix. Rendering test works.
git-svn-id: https://easyrpg.svn.sourceforge.net/svnroot/easyrpg/trunk@538 c7c03aae-3e16-0410-b5dd-fde515843ac6

+++ editor/src/FrameEditor.cpp
@@ -15,6 +15,7 @@ You should have received a copy of the GNU General Public License
 along with this program.If not, see <http://www.gnu.org/licenses/>.*/
 
 #include <wx/wx.h>
+#include <iostream>
 #include "FrameEditor.h"
 #include "DialogMaterial.h"
 
@@ -489,7 +490,7 @@ void ScrolledCanvas::OnDraw(wxDC& dc)
 
 void ScrolledCanvas::SetScale(float zoom){
 	Scale = zoom;
-	this->SetScrollbars(16*zoom, 16*zoom, this->m_data.MapWidth, this->m_data.MapHeight, 0, 0);
+	this->SetScrollbars(16*(int)zoom, 16*(int)zoom, this->m_data.MapWidth, this->m_data.MapHeight, 0, 0);
 }
 
 bool ScrolledCanvas::load_map(wxString FileName)
